WK8_Monitoring_2/code_otel/src/requirements.txt
Flask>=1.1.2
opentelemetry-api>=1.12.0
opentelemetry-sdk>=1.12.0
opentelemetry-exporter-prometheus>=0.44b0
uwsgi>=2.0.18
